Moved batched poseidon related functions to a separate file
