+++ primitives/Cargo.toml
@@ -77,3 +77,8 @@ required-features = ["vrf"]
 name = "poseidon_crh"
 path = "benches/crypto_primitives/poseidon_crh.rs"
 harness = false
+
+[[bench]]
+name = "poseidon_mht"
+path = "benches/crypto_primitives/poseidon_mht.rs"
+harness = false
